refactor: removed 'datasource' from package path to simplify structure
