--- app/src/main/java/com/sanaa/tudee_assistant/data/datasource/local/dao/CategoryDao.kt
@@ -1,8 +1,0 @@
-package com.sanaa.tudee_assistant.data.datasource.local.dao
-
-import androidx.room.Dao
-
-@Dao
-interface CategoryDao {
-
-}
