+++ app/src/main/java/com/sanaa/tudee_assistant/data/local/dao/CategoryDao.kt
@@ -1,0 +1,8 @@
+package com.sanaa.tudee_assistant.data.local.dao
+
+import androidx.room.Dao
+
+@Dao
+interface CategoryDao {
+
+}
