+++ app/src/main/java/com/sanaa/tudee_assistant/data/local/dao/TaskDao.kt
@@ -1,0 +1,11 @@
+package com.sanaa.tudee_assistant.data.local.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import com.sanaa.tudee_assistant.data.local.entity.TaskEntity
+
+@Dao
+interface TaskDao {
+    @Insert
+    fun insert(task: TaskEntity)
+}
